backend/requirements.txt
Django==3.2
djangorestframework==3.13.0
psycopg2==2.9.4
Markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
PyYAML==6.0
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
